fix: restructured components & cleanup
