+++ packages/frontend/src/retro/components/TimePicker.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { FlexBox } from "./FlexBox";
-import { TextInput } from "./TextInput";
-import IncrementTimerButton from "../../retro/components/buttons/IncrementTimerButton";
+import { FlexBox } from "../../common/components/FlexBox";
+import { TextInput } from "../../common/components/TextInput";
+import IncrementTimerButton from "./buttons/IncrementTimerButton";
 interface TimePickerProps {
   minutes: string;
   seconds: string;
